fix(self-gate): shrink the three files this branch pushed past the large-file floor
Our own self-guardrail blocked the branch: command-defs.ts (516),
create.ts (510), and stop-gate.ts (507) crossed the 500-line large-file
threshold with this release's additions — net-new medium findings vs
the baseline. Per the house rule (and the PR #134 precedent): fix the
findings, never refresh the baseline. Cohesive extractions, not
comment-trimming:

- create.ts → envelope-meta.ts: the repo/analysis-environment metadata
  derivation (hashContent, readRepoState, buildAnalysisMeta) — the
  envelope hashes are their own concern, distinct from finding identity.
- stop-gate.ts → stop-gate-io.ts: the hook-boundary plumbing
  (StopHookPayload, readStdinPayload, runConfiguredTests) — the gate
  module keeps only decision logic.
- command-defs.ts → command-defs-internal.ts: the internal-audience
  registry partition, spread back into COMMANDS so the one registry and
  the literal CommandId union are unchanged.

Also suppress the slop-gate hit on the floor-debt test fixture: the
console.error lives inside the fixture's OWN fake-compiler script
string (it IS the simulated build error), annotated slop-ok. Lesson
recorded: the slop range-diff only sees committed changes — run it
after committing, not before.
